requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.1
pdfplumber==0.10.3
python-docx==1.0.1
